% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability Density Function of Discrete
%   Distributions.R
\name{pdBinom}
\alias{pdBinom}
\title{Probability Density Function of Binomial Distribution}
\usage{
pdBinom(x, n, p)
}
\arguments{
\item{x}{is the random variable}

\item{n}{is the number of samples}

\item{p}{is the success probability of each trial}
}
\value{
the probability density function of binomial distribution
}
\description{
Compute the probability density function of binomial distribution
}
\examples{
pdBinom1 <- pdBinom(1, 5, 0.6);
pdBinom2 <- pdBinom(6, 10, 0.25);
}
