% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability Density Function of Discrete
%   Distributions.R
\name{pdGeom}
\alias{pdGeom}
\title{Probability Density Function of Geometric Distribution}
\usage{
pdGeom(x, p)
}
\arguments{
\item{x}{is the random variable}

\item{p}{is the success probability of each trial}
}
\value{
The probability density function of geometric distribution
}
\description{
Compute probability density function of geometric distribution
}
\examples{
pdGeom1 <- pdGeom(5, 0.01);
pdGeom2 <- pdGeom(3, 0.8);
}
