% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability Density Function of Discrete
%   Distributions.R
\name{pdNB}
\alias{pdNB}
\title{Probability Density Function of Negative Binomial Distribution}
\usage{
pdNB(x, r, p)
}
\arguments{
\item{x}{is the random variable}

\item{r}{is the number of successes}

\item{p}{is the success probability of each trial}
}
\value{
The probability density function of negative binomial distribution
}
\description{
Compute probability density function of negative binomial distribution
}
\examples{
pdNB1 <- pdNB(15, 5, 0.2);
pdNB2 <- pdNB(5, 2, 0.25);
}
